# frontend/openapi3.json
{
    "openapi": "3.0.0",
    "info": {
        "description": "Booking and appointment management API.",
        "title": "Fejd API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/admin/business/{businessID}/employees/{userID}/image": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Uploads an employee avatar (multipart file). Visibility is public so avatars render on the public barbers page.",
                "tags": [
                    "admin"
                ],
                "summary": "Upload an employee avatar",
                "parameters": [
                    {
                        "description": "Business UUID",
                        "name": "businessID",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "User ID (Keycloak sub)",
                        "name": "userID",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/postApiAdminBusiness_businessid_employees_userid_image"
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/dto.Image"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handler.ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handler.ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handler.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/admin/business/{businessID}/images": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Uploads a business hero/logo/background image (multipart file). Visibility is public.",
                "tags": [
                    "admin"
                ],
                "summary": "Upload a business image",
                "parameters": [
                    {
                        "description": "Business UUID",
                        "name": "businessID",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "purpose": {
                                        "description": "hero|logo|background (default hero)",
                                        "type": "string"
                                    },
                                    "file": {
                                        "description": "Image file",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "required": [
                                    "file"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/dto.Image"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handler.ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handler.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/admin/business/{businessID}/images/{imageID}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes an image. Admins delete the whole image; an employee may remove only their own avatar link.",
                "tags": [
                    "admin"
                ],
                "summary": "Delete an image",
                "parameters": [
                    {
                        "description": "Business UUID",
                        "name": "businessID",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Image UUID",
                        "name": "imageID",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handler.MessageResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handler.ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handler.ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handler.ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handler.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/admin/business/{businessID}/services/{serviceID}/image": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Uploads a service picture (multipart file). Visibility is public; services.picture_id is updated.",
                "tags": [
                    "admin"
                ],
                "summary": "Upload a service picture",
                "parameters": [
                    {
                        "description": "Business UUID",
                        "name": "businessID",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Service UUID",
                        "name": "serviceID",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/postApiAdminBusiness_businessid_employees_userid_image"
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/dto.Image"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handler.ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handler.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/images/{imageID}": {
            "get": {
                "description": "Public images are served directly; private images require auth and access (streamed through the API).",
                "tags": [
                    "public"
                ],
                "summary": "Retrieve an image",
                "parameters": [
                    {
                        "description": "Image UUID",
                        "name": "imageID",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "image/*": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "image/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/handler.ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "image/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/handler.ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "image/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/handler.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "servers": [
        {
            "url": "//localhost:8080"
        }
    ],
    "components": {
        "requestBodies": {
            "postApiAdminBusiness_businessid_employees_userid_image": {
                "content": {
                    "multipart/form-data": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "file": {
                                    "description": "Image file",
                                    "type": "string",
                                    "format": "binary"
                                }
                            },
                            "required": [
                                "file"
                            ]
                        }
                    }
                },
                "required": true
            }
        },
        "securitySchemes": {
            "BearerAuth": {
                "description": "Type \"Bearer\" followed by a space and JWT token.",
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        },
        "schemas": {
...
            "dto.Image": {
                "type": "object",
                "required": [
                    "created_at",
                    "id",
                    "url"
                ],
                "properties": {
                    "content_type": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "id": {
                        "type": "string"
                    },
                    "url": {
                        "type": "string"
                    }
                }
            },
...
            "handler.ErrorResponse": {
                "type": "object",
                "required": [
                    "error"
                ],
                "properties": {
                    "error": {
                        "type": "string",
                        "example": "error message"
                    }
                }
            },
            "handler.MessageResponse": {
                "type": "object",
                "required": [
                    "message"
                ],
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "operation complete"
                    }
                }
            },
...
        }
    }
}